requirements.txt
flask==3.0.2
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.2.0
pywebview==5.0.6
python-dotenv==1.0.1
opencv-python==4.11.0.86
numpy==2.2.6
pyzbar==0.1.9
scikit-learn==1.5.0
transformers==4.51.3
torch==2.2.2
sentencepiece==0.2.0
pandas==2.2.2
PyMuPDF==1.25.5 